Add sqlalchemy migration
